replace `bincode` with `serde_json` for moss-db implementation
`bincode` serializes the data in a non self-describing way: this means that it cannot be used to serialize complex Rust enums with recursion. `serde_json` will serialize the metadata alongside with the data, thus it can handle such types with no issue. We are still storing binary representation in the database, which means that there should only be a modest increase in data size, reflecting the overhead from JSON format, compared to `bincode`.