Demo for emitting log entry to the frontend
Implement a basic demo for how the `LoggingService` will emit log events to the frontend. The `LogsPage` will listen to `logging` channel for incoming LogEntry, and a test button will call a tauri command that generates a test log using the `LoggingService`, which will emit a event back to the frontend.