Refactor the logic for finding which files to delete logs from
Previously, we were using a rather complex logic to find which finds to delete logs from: we would find the start and end timestamp of the entries to be deleted, find files that cover that range, and process each file. Since we are now storing the path of each log, we can easily narrow down the list of files we need to update for a bulk deletion task.