Spawn blocking thread for repo related operations
There are some issues trying to integrate `git2rs` into our application, since it's fundamentally sync. This sync nature forces us to use a blocking version of `oauth2` and `reqwest`, which seems to be creating a tokio runtime under the hood. I think the best strategy is to spawn a blocking thread for git repo related operations, instead of trying to fit `git2rs` into async.